fix: restart the client SSE connection if it fails
A previous change would restart the client SSE connection if the
connection was closed. This change will check whether the connection
needs to be restarted when another message comes through. Without this
change, an SSE client would be useless if the connection errored for
some reason. No harm in restarting it.

Signed-off-by: Donnie Adams <donnie@acorn.io>