fix: handle empty initialized response
An MCP server can respond with content types application/json or
text/event-stream. Typically, if a server responds with
text/event-stream, then the content length is -1. However, if the
response body is empty with no content type, then the content length
could still be -1.

This change better detects when we should read a response.

Signed-off-by: Donnie Adams <donnie@acorn.io>